Move GenericRecord interfaces to its own package
The team decided to move the GenericRecord related interfaces
to its own package.

This PR moves the GenericRecord and GenericRecordBuilder interfaces
to the new `serialization.genericrecord` package.
