+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/inject/PortableUpsertTarget.java
@@ -20,7 +20,7 @@ import com.hazelcast.internal.serialization.impl.portable.PortableGenericRecordB
 import com.hazelcast.nio.serialization.ClassDefinition;
 import com.hazelcast.nio.serialization.FieldDefinition;
 import com.hazelcast.nio.serialization.FieldType;
-import com.hazelcast.nio.serialization.GenericRecord;
+import com.hazelcast.nio.serialization.genericrecord.GenericRecord;
 import com.hazelcast.sql.impl.QueryException;
 import com.hazelcast.sql.impl.type.QueryDataType;
