+++ hazelcast/src/main/java/com/hazelcast/nio/serialization/FieldKind.java
@@ -16,6 +16,7 @@
 
 package com.hazelcast.nio.serialization;
 
+import com.hazelcast.nio.serialization.genericrecord.GenericRecord;
 import com.hazelcast.spi.annotation.Beta;
 
 /**
